+++ app/assets/stylesheets/components/_navbar.scss
@@ -1,5 +1,9 @@
 .navbar-lewagon {
   justify-content: space-between;
+  background: white;
+}
+
+.navbar-home {
   background: transparent;
   position: fixed;
   width: 100%;
