+++ app/javascript/packs/application.js
@@ -28,7 +28,7 @@ import "bootstrap";
 // Internal imports, e.g:
 // import { initSelect2 } from '../components/init_select2';
 
-import { navbarScroll } from './components/navbar_scroll';
+import { navbarScroll } from '../components/navbar_scroll';
 
 document.addEventListener('turbolinks:load', () => {
   // Call your functions here, e.g:
