--- app/javascript/packs/components/navbar_scroll.js
@@ -1,9 +1,0 @@
-const navbarScroll = () => {
-  window.addEventListener("scroll", () => {
-    const navbar = document.querySelector(".navbar-lewagon");
-    console.log(navbar);
-    navbar.classList.toggle("scrolled", window.scrollY > window.innerHeight);
-  });
-};
-
-export  { navbarScroll }
